requirements.txt
# AuraProject AI Service v2.4.0 - Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Image Processing
pillow>=10.1.0
numpy>=1.24.0
scipy>=1.11.0  # For morphological operations in mask processing

# LLM
openai>=1.6.0
httpx>=0.25.0
google-generativeai>=0.3.0

# Segmentation & Attributes
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Virtual Try-On
diffusers>=0.24.0
accelerate>=0.25.0

# Database
pymongo>=4.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
